[ERRORS] Sessions must wait for admin approval before being locked
This adds a new value to the union type for LockedStatus. Because
assertNever is setup in the switch statement for locked-icon.tsx,
TypeScript will statically ensure the new case is handled in the switch
statement. It helpfully points out which case is missing, although its
wording is a little strange.

locked-icon.tsx:19:26 - error TS2345: Argument of type '"Pending"' is not assignable to parameter of type 'never'.

19       return assertNever(lockedStatus);
                            ~~~~~~~~~~~~