[ERRORS] LockedStatus is a string now
This demonstrates the importance of enumerating values. Without doing
so a lot of the benefits of TypeScript like exhaustiveness checking and
type compatibility don't work as well as they could.

Here, even though all the possible cases are enumerated, TypeScript
can't know that because it thinks there are infinite possible cases (as
there are infinite strings).

locked-icon.tsx:20:26 - error TS2345: Argument of type 'string' is not assignable to parameter of type 'never'.

20       return assertNever(lockedStatus);